+++ examples/basic.py
@@ -1,19 +1,15 @@
 #!/usr/bin/env python3
 
-from pylecroy.pylecroy import *
-import numpy as np
-import matplotlib.pyplot as plt
+from pylecroy.pylecroy import Lecroy
 import sys
-
-VERSION = '0.1.0'
 
 USAGE = '''basic: basic usage
 Usage:
     python basic.py -a "IP:10.67.16.22"
-
+    python basic.py -a VXI11:10.67.0.211
+    
 Options:
     -h, --help              this help message.
-    -v, --version           version info.
     -a, --address           device IP address
 '''
 
@@ -24,15 +20,12 @@ def main(argv=None):
     if argv is None:
         argv = sys.argv[1:]
     try:
-        opts, args = getopt.gnu_getopt(argv, 'hva:', ['help', 'version', 'address='])
+        opts, args = getopt.gnu_getopt(argv, 'ha:', ['help', 'address='])
         address = None
 
         for o, a in opts:
             if o in ('-h', '--help'):
                 print(USAGE)
-                return 0
-            if o in ('-v', '--version'):
-                print(VERSION)
                 return 0
             elif o in ('-a', '--address'):
                 address = a