+++ tests/tests.py
@@ -1,54 +1,52 @@
 from pylecroy.pylecroy import Lecroy
-from pylecroy.pylecroy import TriggerModes
-from pylecroy.pylecroy import Channels
-from pylecroy.pylecroy import GridStates
-from pylecroy.pylecroy import Memories
-
+from pylecroy.pylecroy import Trigger
+from pylecroy.pylecroy import WaveForm
+from pylecroy.pylecroy import Grid
 import time
 
 ADDRESS_USB = "USBTMC:USB0::0x05ff::0x1023::2816N63170::INSTR"
 ADDRESS_TCP = "IP:10.67.0.35"
-
+ADDRESS_VXI11 = "VXI11:10.67.0.211"
 if __name__ == '__main__':
     # Connection
-    device = Lecroy(ADDRESS_USB)
+    device = Lecroy(ADDRESS_VXI11)
 
     # Get scope identifier
     print("scope identifier     : {} ".format(device.identifier))
 
     # Display channels
     print("Channels  OFF...")
-    for channel in [Channels.C1, Channels.C2, Channels.C3, Channels.C4]:
+    for channel in [WaveForm.Channels.C1, WaveForm.Channels.C2, WaveForm.Channels.C3, WaveForm.Channels.C4]:
         device.display_channel(channel, "OFF")
         time.sleep(0.5)
 
     print("Channels  ON...")
-    for channel in [Channels.C1, Channels.C2, Channels.C3, Channels.C4]:
+    for channel in [WaveForm.Channels.C1, WaveForm.Channels.C2, WaveForm.Channels.C3, WaveForm.Channels.C4]:
         device.display_channel(channel, "ON")
         time.sleep(0.5)
 
     print("Channels  OFF...")
-    for channel in [Channels.C1, Channels.C2, Channels.C3, Channels.C4]:
+    for channel in [WaveForm.Channels.C1, WaveForm.Channels.C2, WaveForm.Channels.C3, WaveForm.Channels.C4]:
         device.display_channel(channel, "OFF")
         time.sleep(0.5)
 
-    device.display_channel(Channels.C1, "ON")
+    device.display_channel(WaveForm.Channels.C1, "ON")
     print("Grid : {0}".format(device.grid))
 
     input("Get a signal on C1 and press a key to continue...")
 
     # grid state
-    for state in GridStates.STATES:
-        print("Grid state : " + state)
+    for state in Grid.States:
+        print(f"Grid state : {state.value}")
         device.grid = state
         time.sleep(0.5)
 
-    device.grid = GridStates.SINGLE
+    device.grid = Grid.States.SINGLE
 
     # Controlling Waveform capture
-    for trigger in TriggerModes.MODES:
+    for trigger in Trigger.Modes:
         device.trigger_mode = trigger
-        print("scope trigger mode   : {}".format(device.trigger_mode))
+        print(f"scope trigger mode   : {device.trigger_mode}")
         time.sleep(0.5)
 
     try:
@@ -56,25 +54,25 @@ if __name__ == '__main__':
     except ValueError as msg:
         print(msg)
 
-    print("scope sequence       : {}".format(device.sequence))
+    print(f"scope sequence       : {device.sequence}")
 
     # HARDCOPY - Printing the Display/Screen Capture
-    print("hardcopy : {0}".format(device.get_hardcopy()))
+    print(f"hardcopy : {device.hardcopy}")
     new_cfg = {'DEV': 'JPEG', 'FORMAT': 'LANDSCAPE'}
-    device.set_hardcopy(new_cfg)
-    print("hardcopy : {0}".format(device.get_hardcopy()))
+    device.hardcopy = new_cfg
+    print(f"hardcopy : {device.hardcopy}")
     device.store_hardcopy("C1")
     new_cfg = {'DEV': 'BMP', 'FORMAT': 'PORTRAIT'}
-    device.set_hardcopy(new_cfg)
-    print("hardcopy : {0}".format(device.get_hardcopy()))
+    device.hardcopy = new_cfg
+    print(f"hardcopy : {device.hardcopy}")
     device.store_hardcopy("C1")
 
     # Preserving and Restoring Waveforms
     print("Set Trigger Single...")
-    device.trigger_mode = TriggerModes.SINGLE
+    device.trigger_mode = Trigger.Modes.SINGLE
     # save C1 to M1
     print("Channel C1 save in M1...")
-    device.save_memory(Channels.C1, Memories.M1)
+    device.save_memory(WaveForm.Channels.C1, WaveForm.Memories.M1)
     print("Show M1...")
-    device.display_channel(Memories.M1, "ON")
+    device.display_channel(WaveForm.Memories.M1, "ON")
     device.close()